Combine test helpers into test
